change price type from Float to BigDecimal

+++ bms-backend/src/main/java/com/humanoo/bms_backend/dto/BookDto.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -14,5 +16,5 @@ public class BookDto {
     private String title;
     private String author;
     private String isbn;
-    private Float price;
+    private BigDecimal price;
 }
